+++ apps/opik-backend/src/test/java/com/comet/opik/api/resources/utils/FeedbackScoreAssertionUtils.java
@@ -1,6 +1,5 @@
 package com.comet.opik.api.resources.utils;
 
-import com.comet.opik.api.Comment;
 import com.comet.opik.api.ExperimentItem;
 import com.comet.opik.api.FeedbackScoreNames;
 import lombok.experimental.UtilityClass;
@@ -10,10 +9,7 @@ import java.util.List;
 import static org.assertj.core.api.Assertions.assertThat;
 
 @UtilityClass
-public class AssertionUtils {
-
-    public static final String[] IGNORED_FIELDS_COMMENTS = {"id", "createdAt", "lastUpdatedAt", "createdBy",
-            "lastUpdatedBy"};
+public class FeedbackScoreAssertionUtils {
 
     public static void assertFeedbackScoreNames(FeedbackScoreNames actual, List<String> expectedNames) {
         assertThat(actual.scores()).hasSize(expectedNames.size());
@@ -46,35 +42,5 @@ public class AssertionUtils {
                                 .build())
                         .toList())
                 .build();
-    }
-
-    public static void assertTraceComment(Comment expected, Comment actual) {
-        assertThat(actual)
-                .usingRecursiveComparison()
-                .ignoringFields(IGNORED_FIELDS_COMMENTS)
-                .isEqualTo(expected);
-
-        assertThat(actual.createdAt()).isNotNull();
-        assertThat(actual.lastUpdatedAt()).isNotNull();
-        assertThat(actual.createdBy()).isNotNull();
-        assertThat(actual.lastUpdatedBy()).isNotNull();
-    }
-
-    public static void assertUpdatedComment(Comment initial, Comment updated, String expectedText) {
-        assertThat(initial.text()).isNotEqualTo(expectedText);
-
-        assertTraceComment(initial.toBuilder().text(expectedText).build(), updated);
-
-        assertThat(updated.createdAt()).isEqualTo(initial.createdAt());
-        assertThat(updated.lastUpdatedAt()).isNotEqualTo(initial.lastUpdatedAt());
-        assertThat(updated.createdBy()).isEqualTo(initial.createdBy());
-        assertThat(updated.lastUpdatedBy()).isEqualTo(initial.lastUpdatedBy());
-    }
-
-    public static void assertComments(List<Comment> expected, List<Comment> actual) {
-        assertThat(actual)
-                .usingRecursiveComparison()
-                .ignoringFields(IGNORED_FIELDS_COMMENTS)
-                .isEqualTo(expected);
     }
 }
